+++ app/pages/Account/Tabs/ModulesTab/Contract.tsx
@@ -406,6 +406,7 @@ function Contract({
                 <Code
                   sourceBytecode={selectedModule?.source}
                   moduleBytecode={selectedModuleBytecode?.bytecode}
+                  codeLinkContext={{packageAddress: address, isObject}}
                 />
               </Box>
             </>